src/main/java/com/chairmo/controller/CustomerController.java
package com.chairmo.controller;

import com.chairmo.exception.ObjectNotFoundException;
import com.chairmo.model.Customer;
import com.chairmo.repositories.services.CustomerService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Optional;

@RestController
@RequestMapping("/customers")
public class CustomerController {
    private CustomerService customerService;

    @Autowired
    public void setCustomerService(CustomerService customerService) {
        this.customerService = customerService;
    }

    @GetMapping(path = "/{id}", produces = "application/json")
    public ResponseEntity<Optional<Customer>> getCustomer(@PathVariable(value = "id") Integer id) {
        return new ResponseEntity<>(customerService.findById(id), HttpStatus.OK);
    }

    @GetMapping(path = "/", produces = "application/json")
    public ResponseEntity<List<Customer>> getAllCustomers() {
        return new ResponseEntity<>(customerService.findAll(), HttpStatus.OK);
    }

    @PostMapping("/new")
    public ResponseEntity<Customer> createCustomer(@RequestBody Customer customer) {
        return new ResponseEntity<>(customerService.save(customer), HttpStatus.CREATED);
    }

    @PutMapping("/{id}")
    public ResponseEntity<Customer> updateCustomer(@PathVariable Integer id, @RequestBody Customer customer) {
        Optional<Customer> customer1 = customerService.findById(id);
        if (!customer1.isPresent()) throw new ObjectNotFoundException("id not found");
        else
        return new ResponseEntity<>(customerService.save(customer), HttpStatus.OK);
    }

    @DeleteMapping("/{id}")
    public void deleteCustomer(@PathVariable Integer id) {
        if (id != null) {
            customerService.deleteById(id);
        } else {
            throw new ObjectNotFoundException("id not found");
        }
    }
}
